+++ tools/plugins/acf-extensions/acf-extensions.php
@@ -1,8 +1,8 @@
 <?php
 /*
-    Plugin Name: Advanced Custom Fields: Component Field
-    Description: Advanced Custom Fields add on. Make an entire acf field group reuseable, as a component field.
-    Version: 2.0.0
+    Plugin Name: Advanced Custom Fields Extensions
+    Description: Advanced Custom Fields add on. Create component, components field, hidden field and lastest post field
+    Version: 1.0.0
     Author: Metabolism
     License: GPLv2 or later
     License URI: http://www.gnu.org/licenses/gpl-2.0.html
@@ -16,10 +16,12 @@ defined('ABSPATH') or die('No script kiddies please!');
  * @since  1.0.4
  * @return void
  */
-function acf5_component_field_load_textdomain() {
+function acf_extensions_load_textdomain() {
     load_plugin_textdomain('acf-component_field', false, dirname(plugin_basename(__FILE__)).'/lang/');
 }
-add_action('plugins_loaded', 'acf5_component_field_load_textdomain');
+
+add_action('plugins_loaded', 'acf_extensions_load_textdomain');
+
 
 /**
  * Load the main field class after core acf init
@@ -27,11 +29,19 @@ add_action('plugins_loaded', 'acf5_component_field_load_textdomain');
  * @since  1.0.0
  * @return void
  */
-function include_acf5_component_field_plugin() {
-	include_once('acf-component_field-v5.php');
-    $GLOBALS['acf_component_field'] = new acf_field_component_field();
+function include_acf_extensions_plugin() {
+
+	include_once('acf-component_field.php');
+	include_once('acf-hidden_field.php');
+	include_once('acf-latest_posts_field.php');
+
+	acf_register_field_type( 'acf_field_component' );
+	acf_register_field_type( 'acf_field_hidden' );
+	acf_register_field_type( 'acf_field_latest_posts' );
 }
-add_action('acf/include_field_types', 'include_acf5_component_field_plugin');
+
+add_action('acf/include_field_types', 'include_acf_extensions_plugin');
+
 
 /**
  * Change acf-component status back to acf-diabled when deactivated
@@ -39,9 +49,9 @@ add_action('acf/include_field_types', 'include_acf5_component_field_plugin');
  * @since  1.0.0
  * @return void
  */
-function deactivate_acf5_component_field_plugin() {
+function deactivate_acf_extensions_plugin() {
 
-    // conver all the acf-comopnent to acf-disabled
+    // convert all the acf-component to acf-disabled
     $args = array(
         'posts_per_page' => -1,
         'post_type'      => 'acf-field-group',
@@ -55,10 +65,12 @@ function deactivate_acf5_component_field_plugin() {
         wp_update_post($field_group);
     }
 }
-register_deactivation_hook(__FILE__, 'deactivate_acf5_component_field_plugin');
+
+register_deactivation_hook(__FILE__, 'deactivate_acf_extensions_plugin');
+
 
 /**
- * Loop throught the inactive acf field group and tag them as acf component
+ * Loop through the inactive acf field group and tag them as acf component
  * don't touch the enable ones even its is_acf_component is true.
  * Because the means user manually enable it,
  * thus might not be a component anymore
@@ -66,7 +78,7 @@ register_deactivation_hook(__FILE__, 'deactivate_acf5_component_field_plugin');
  * @since  1.0.7
  * @return void
  */
-function activate_acf5_component_field_plugin() {
+function activate_acf_extensions_plugin() {
 
     $args = array(
         'posts_per_page' => -1,
@@ -87,4 +99,5 @@ function activate_acf5_component_field_plugin() {
         }
     }
 }
-register_activation_hook(__FILE__, 'activate_acf5_component_field_plugin');
+
+register_activation_hook(__FILE__, 'activate_acf_extensions_plugin');