+++ changelog.md
@@ -4,7 +4,8 @@ CHANGELOG
 ##4.0.1
 ####Bugfix
 - menu children not showing up in tree
-
+####Added
+- ACF Extensions plugin with components, hidden field ans lastest posts
 
 ##4.0.0
 first official release fr SF 4